requirements.txt
numpy==1.24.2
pyserial==3.5
